Hola  ∾ prgtest autocommit 20_03 #4

+++ 20_03_csv/Mitjana.java
@@ -28,7 +28,7 @@ public class Mitjana {
                     continue;
                 total += Integer.parseInt(value);
             }
-            int media = total/splitByComma.length - 1;
+            float media = total/splitByComma.length - 1;
 
             System.out.println(splitByComma[0] + " (" + media + "," + total + ")");
         }
